src/main/java/softLaunch/controller/ClientController.java
package softLaunch.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import softLaunch.domain.Client;
import softLaunch.service.ClientService;
import softLaunch.domain.RequestWrapper;

import java.util.List;

@RestController
@RequestMapping("/client")
public class ClientController {

    @Autowired
    private ClientService clientService;

    @GetMapping
    @ResponseStatus(HttpStatus.OK)
    public List<Client> getAllClients() {
        return clientService.getAllClients();
    }

    @PostMapping("/batch")
    @ResponseStatus(HttpStatus.CREATED)
    public ResponseEntity<RequestWrapper> postWithMultipleObjects(@RequestBody RequestWrapper requestWrapper){
        return clientService.addBatch(requestWrapper);
    }
}
